+++ cmake/hardrt_tester.cmake
@@ -29,7 +29,11 @@ if(HARDRT_PORT STREQUAL "posix")
           ${CMAKE_SOURCE_DIR}/tests/test_sleep_zero.c
           ${CMAKE_SOURCE_DIR}/tests/test_task_return.c
           ${CMAKE_SOURCE_DIR}/tests/test_semaphore.c
+          ${CMAKE_SOURCE_DIR}/tests/test_queue.c
           ${CMAKE_SOURCE_DIR}/tests/test_external_tick.c
+          ${CMAKE_SOURCE_DIR}/tests/test_idle_behavior.c
+          ${CMAKE_SOURCE_DIR}/tests/test_mutex.c
+          ${CMAKE_SOURCE_DIR}/tests/test_now_ms.c
   )
 
   target_link_libraries(hardrt_tests PRIVATE ${LIB_NAME})